pyproject.toml
[project]
name = "llm-with-web-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.9.2",
    "fastmcp>=2.13.2",
    "pydantic-ai>=1.26.0",
    "python-dotenv>=1.2.1",
]
